requirements.txt
# ONNX Profiling Environment Requirements
# Generated for Vision Transformer benchmarking

# Core ONNX packages
onnxruntime-gpu==1.19.2
onnx==1.19.1

# Data processing
numpy>=2.0.0
pandas>=2.2.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0

# System monitoring
psutil>=7.0.0

# Utilities
tqdm>=4.67.0

